refactor: move commit control selection

AI-Model: Qoder Auto
AI-Contributed/Feature: 40/40
AI-Contributed/UT: 0/0
